+++ force-app/main/default/classes/AccountTriggerTest.cls
@@ -8,7 +8,7 @@ private class AccountTriggerTest {
 
     // Inserindo a conta
     insert acc;
-    System.debug('ULTIMO TEaaaSTaaabbbbbE');
+    System.debug('ULTIMO TEaaaSTaaaaaE');
 
     // Recuperando a conta inserida do banco de dados
     Account insertedAccount = [
@@ -17,8 +17,8 @@ private class AccountTriggerTest {
       WHERE Id = :acc.Id
     ];
 
-    // new testClass();
-    // new Test();
+    new testClass();
+    new Test();
 
     // Verificando se a trigger atualizou corretamente os campos esperados
     // System.assertEquals('Test Account', insertedAccount.Name);
